move more recognizer logic into topography sublibrary
